BE_unified_state/backend/requirements.txt
pandas
openpyxl
PyYAML
python-hcl2
numpy